feat: added hashmap header fiel
